summarizationproject/requirements.txt
asgiref==3.4.1
blis==0.7.5
catalogue==2.0.6
certifi==2020.6.20
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cymem==2.0.6
Django==4.0.1
djangorestframework==3.13.1
idna==3.3
Jinja2==3.0.3
joblib==1.1.0
langcodes==3.3.0
lxml==4.7.1
MarkupSafe==2.0.1
murmurhash==1.0.6
mysqlclient==2.1.0
nltk==3.6.7
numpy==1.22.1
packaging==21.3
pathy==0.6.1
preshed==3.0.6
pydantic==1.8.2
pyparsing==3.0.6
python-docx==0.8.11
pytz==2021.3
regex==2022.1.18
requests==2.27.1
rouge==1.0.1
simplejson==3.17.6
six==1.16.0
smart-open==5.2.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
sqlparse==0.4.2
srsly==2.4.2
thinc==8.0.13
tqdm==4.62.3
typer==0.4.0
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.8
wasabi==0.9.0
wincertstore==0.2
